Remove check for whether a pomodoro completes a task.
It's not particularly likely that a task will be completed exactly
at the end of a pomodoro, and it's only one command to mark it as
done, anyway.

Also remove the connection to type breaks, which I'm no longer using,
and isn't needed anyway now it always goes to OPEN at the end of a
pomodoro.